# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Dhasvin Enterprises Inventory Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user info
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
  /bills:
    post:
      operationId: createBill
      tags: [bills]
      summary: Generate a bill/invoice
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBillRequest"
      responses:
        "201":
          description: Bill created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bill"
    get:
      operationId: getBills
      tags: [bills]
      summary: Get all bills
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of bills
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Bill"
  /bills/{id}:
    get:
      operationId: getBill
      tags: [bills]
      summary: Get a bill by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Bill details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bill"
  /analytics/dashboard:
    get:
      operationId: getDashboardStats
      tags: [analytics]
      summary: Get dashboard KPI stats
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Dashboard statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardStats"
  /analytics/sales-trend:
    get:
      operationId: getSalesTrend
      tags: [analytics]
      summary: Get sales trend over time (daily/monthly)
      security:
        - bearerAuth: []
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [daily, monthly]
        - name: startDate
          in: query
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Sales trend data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TrendDataPoint"
  /analytics/top-products:
    get:
      operationId: getTopProducts
      tags: [analytics]
      summary: Get top selling products
      security:
        - bearerAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Top products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TopProduct"
  /analytics/low-stock:
    get:
      operationId: getLowStockProducts
      tags: [analytics]
      summary: Get low stock alerts
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Low stock products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LowStockAlert"
  /analytics/category-breakdown:
    get:
      operationId: getCategoryBreakdown
      tags: [analytics]
      summary: Get inventory breakdown by category
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Category breakdown data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CategoryBreakdown"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
...
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required:
        - email
        - password
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required:
        - token
        - user
    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [admin, staff, user]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - email
        - name
        - role
...
    BillItem:
      type: object
      properties:
        productId:
          type: integer
        productName:
          type: string
        hsnCode:
          type: string
        quantity:
          type: integer
        unitPrice:
          type: number
          format: float
        discount:
          type: number
          format: float
        gst:
          type: number
          format: float
        amount:
          type: number
          format: float
      required:
        - productId
        - quantity
        - unitPrice
        - amount
    Bill:
      type: object
      properties:
        id:
          type: integer
        billNumber:
          type: string
        customerName:
          type: string
        customerAddress:
          type: string
        customerPhone:
          type: string
        customerGst:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/BillItem"
        subtotal:
          type: number
          format: float
        totalDiscount:
          type: number
          format: float
        totalGst:
          type: number
          format: float
        totalAmount:
          type: number
          format: float
        amountInWords:
          type: string
        notes:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - billNumber
        - items
        - totalAmount
    CreateBillRequest:
      type: object
      properties:
        customerName:
          type: string
        customerAddress:
          type: string
        customerPhone:
          type: string
        customerGst:
          type: string
        items:
          type: array
          items:
            type: object
            properties:
              productId:
                type: integer
              quantity:
                type: integer
              unitPrice:
                type: number
                format: float
              discount:
                type: number
                format: float
              gst:
                type: number
                format: float
            required:
              - productId
              - quantity
              - unitPrice
        notes:
          type: string
      required:
        - customerName
        - items
...
    DashboardStats:
      type: object
      properties:
        totalProducts:
          type: integer
        totalCategories:
          type: integer
        totalSuppliers:
          type: integer
        totalCustomers:
          type: integer
        pendingInvoices:
          type: integer
        totalSalesToday:
          type: number
          format: float
        totalSalesMonth:
          type: number
          format: float
        monthlyRevenue:
          type: number
          format: float
        totalRevenue:
          type: number
          format: float
        lowStockCount:
          type: integer
        outOfStockCount:
          type: integer
        totalInventoryValue:
          type: number
          format: float
        recentSalesCount:
          type: integer
      required:
        - totalProducts
        - totalSalesToday
        - totalSalesMonth
        - totalRevenue
        - lowStockCount
        - totalCustomers
        - pendingInvoices
        - monthlyRevenue
    TrendDataPoint:
      type: object
      properties:
        date:
          type: string
        sales:
          type: number
          format: float
        purchases:
          type: number
          format: float
        count:
          type: integer
      required:
        - date
        - sales
    TopProduct:
      type: object
      properties:
        productId:
          type: integer
        productName:
          type: string
        totalQuantitySold:
          type: integer
        totalRevenue:
          type: number
          format: float
        categoryName:
          type: string
      required:
        - productId
        - productName
        - totalQuantitySold
        - totalRevenue
    LowStockAlert:
      type: object
      properties:
        productId:
          type: integer
        productName:
          type: string
        quantity:
          type: integer
        minQuantity:
          type: integer
        categoryName:
          type: string
        supplierName:
          type: string
        status:
          type: string
          enum: [low, out_of_stock]
      required:
        - productId
        - productName
        - quantity
        - status
    CategoryBreakdown:
      type: object
      properties:
        categoryId:
          type: integer
        categoryName:
          type: string
        productCount:
          type: integer
        totalValue:
          type: number
          format: float
        totalQuantity:
          type: integer
      required:
        - categoryName
        - productCount
        - totalValue
...